Version 6.9.0

+++ app/models/patient_import.rb
@@ -136,28 +136,33 @@ class PatientImport < ApplicationRecord
       parent_relationships =
         self.parent_relationships.includes(:parent, :patient)
 
-      parents_to_check = parent_relationships.map(&:parent)
-
-      parent_relationships.destroy_all
-
-      parents_to_check.each do |parent|
-        if parent.parent_relationships.empty? && parent.consents.empty?
-          parent.destroy!
-        end
-      end
-
       timestamp = Time.current.to_fs(:long)
       user_name = current_user&.full_name || "system"
       invalidation_note =
         "Consent invalidated on #{timestamp} " \
           "because #{user_name} removed all parent-child relationships from an import."
 
-      Consent.where(id: consents.map(&:id)).update_all(
+      # Invalidate consents first before destroying parent relationships, which
+      # will make consents return an empty set.
+      consents.update_all(
         notes: invalidation_note,
         invalidated_at: Time.current
       )
 
-      StatusUpdaterJob.perform_bulk(consents.map(&:patient_id).uniq.zip)
+      patient_ids_to_update = consents.map(&:patient_id).uniq
+
+      parents_to_check = parent_relationships.map(&:parent)
+
+      parent_relationships.destroy_all
+
+      parents_to_check.each do |parent|
+        if !parent.destroyed? && parent.parent_relationships.empty? &&
+             parent.consents.empty?
+          parent.destroy!
+        end
+      end
+
+      StatusUpdaterJob.perform_bulk(patient_ids_to_update.zip)
     end
   end
 
@@ -172,7 +177,8 @@ class PatientImport < ApplicationRecord
       parent_relationships_without_consents.each(&:destroy!)
 
       parents_to_check.each do |parent|
-        if parent.parent_relationships.empty? && parent.consents.empty?
+        if !parent.destroyed? && parent.parent_relationships.empty? &&
+             parent.consents.empty?
           parent.destroy!
         end
       end
